[bindless] Add core filterability overloads.
Add the needed texture overloads for filterability to the core def file.

Bug: 479566307
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/289855
Commit-Queue: dan sinclair <dsinclair@chromium.org>

+++ src/tint/lang/core/intrinsic/table_data.h
@@ -116,7 +116,7 @@ struct TableIndex {
 };
 
 /// Index type used to index TableData::template_types
-using TemplateIndex = TableIndex<TableIndexNamespace::kTemplate, uint8_t>;
+using TemplateIndex = TableIndex<TableIndexNamespace::kTemplate, uint16_t>;
 
 /// Index type used to index TableData::type_matchers or TableData::number_matchers
 using MatcherIndex = TableIndex<TableIndexNamespace::kMatcher, uint8_t>;
